Moved org.spoutcraft.client.gui.shortcut to org.spoutcraft.client.gui.controls
