+++ src/minecraft/org/spoutcraft/client/gui/controls/GuiCommandsSlot.java
@@ -23,7 +23,7 @@
  * License and see <http://www.spout.org/SpoutDevLicenseV1.txt> for the full license,
  * including the MIT license.
  */
-package org.spoutcraft.client.gui.shortcuts;
+package org.spoutcraft.client.gui.controls;
 
 import net.minecraft.src.FontRenderer;
